pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "wheel"]
build-backend = "setuptools.build_meta"

[versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "rembg/_version.py"
versionfile_build = "rembg/_version.py"
tag_prefix = "v"
parentdir_prefix = "rembg-"

[tool.poetry]
name = "rembg-project"  # Renamed to avoid conflict with dependency
version = "0.1.0"  # Adjust if needed
description = "A Python library for background removal"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"  # Adjusted Python version to include 3.11
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
rembg = "^2.0.61"  # Updated to a more stable version
Pillow = "^9.0.0"  # Add this line to install Pillow
onnxruntime = "^1.15.0"  # Added onnxruntime to resolve missing module error
python-multipart = "*"
gunicorn = "^21.2.0"

[tool.poetry.dev-dependencies]
# Development dependencies (e.g. testing tools like pytest)
# pytest = "^6.2"

[tool.poetry.scripts]
start = "sh -c 'uvicorn main:app --host 0.0.0.0 --port ${PORT:-8000}'"
